refactor: remove EventEmitter from facade
As of right now each module has its own copy of the EventEmitter
contributing to bloat.
