+++ modules/@angular/core/src/event_emitter.ts
@@ -8,9 +8,6 @@
 
 import {Subject} from 'rxjs/Subject';
 
-export {Observable} from 'rxjs/Observable';
-export {Subject} from 'rxjs/Subject';
-
 /**
  * Use by directives and components to emit custom Events.
  *
