+++ modules/@angular/core/test/debug/debug_node_spec.ts
@@ -7,14 +7,12 @@
  */
 
 
-import {Injectable, NO_ERRORS_SCHEMA} from '@angular/core';
+import {EventEmitter, Injectable, NO_ERRORS_SCHEMA} from '@angular/core';
 import {Component, Directive, Input} from '@angular/core/src/metadata';
 import {ComponentFixture, TestBed, async} from '@angular/core/testing';
 import {By} from '@angular/platform-browser/src/dom/debug/by';
 import {getDOM} from '@angular/platform-browser/src/dom/dom_adapter';
 import {expect} from '@angular/platform-browser/testing/matchers';
-
-import {EventEmitter} from '../../src/facade/async';
 
 @Injectable()
 class Logger {
