+++ modules/@angular/core/test/event_emitter_spec.ts
@@ -7,7 +7,7 @@
  */
 
 import {AsyncTestCompleter, beforeEach, describe, expect, inject, it} from '@angular/core/testing/testing_internal';
-import {EventEmitter} from '../src/async';
+import {EventEmitter} from '../src/event_emitter';
 
 export function main() {
   describe('EventEmitter', () => {
