+++ modules/@angular/forms/src/directives/abstract_control_directive.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Observable} from '../facade/async';
+import {Observable} from 'rxjs/Observable';
 import {AbstractControl} from '../model';
 
 /**
