+++ modules/@angular/platform-webworker/src/web_workers/ui/event_dispatcher.ts
@@ -5,7 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {EventEmitter} from '../../facade/async';
+import {EventEmitter} from '@angular/core';
 import {Serializer, SerializerTypes} from '../shared/serializer';
 
 import {serializeEventWithTarget, serializeGenericEvent, serializeKeyboardEvent, serializeMouseEvent, serializeTransitionEvent} from './event_serializer';
