+++ modules/@angular/platform-webworker/src/web_workers/ui/platform_location.ts
@@ -7,10 +7,8 @@
  */
 
 import {LocationChangeListener} from '@angular/common';
-import {Injectable} from '@angular/core';
+import {EventEmitter, Injectable} from '@angular/core';
 import {ɵBrowserPlatformLocation as BrowserPlatformLocation} from '@angular/platform-browser';
-
-import {EventEmitter} from '../../facade/async';
 import {MessageBus} from '../shared/message_bus';
 import {ROUTER_CHANNEL} from '../shared/messaging_api';
 import {LocationType, Serializer, SerializerTypes} from '../shared/serializer';
